Stash splitting into a library
