+++ app/build.gradle
@@ -26,7 +26,7 @@ android {
             //packagingOptions {
             //    doNotStrip '**/*.so'
             //}
-            //debuggable true
+            debuggable true
         }
     }
     compileOptions {
@@ -39,4 +39,5 @@ android {
 dependencies {
     implementation 'androidx.appcompat:appcompat:1.2.0'
     implementation 'androidx.core:core:1.5.0'
+    implementation project(":library")
 }
