+++ app/src/main/java/org/linebender/android/viewdemo/DemoView.java
@@ -2,7 +2,7 @@ package org.linebender.android.viewdemo;
 
 import android.content.Context;
 
-import org.linebender.android.RustView;
+import org.linebender.android.rustview.RustView;
 
 public final class DemoView extends RustView {
     @Override
