+++ library/build.gradle
@@ -1,0 +1,18 @@
+plugins {
+    id 'com.android.library'
+}
+
+android {
+    compileSdk 31
+    defaultConfig {
+        minSdk 28
+        versionCode 1
+        versionName "1.0.0"
+    }
+    namespace 'org.linebender.android.rustview'
+}
+
+dependencies {
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.core:core:1.5.0'
+}
